Fixing doc comments and file structure of NetworkServiceTests
