+++ Tests/NetworkServiceTests/MockURLProtocol.swift
@@ -4,33 +4,32 @@ import Foundation
 class MockURLProtocol: URLProtocol {
     /// Handles incoming requests and returns mocked responses.
     static var requestHandler: ((URLRequest) throws -> (HTTPURLResponse, Data?))?
-
+    
     override class func canInit(with _: URLRequest) -> Bool {
         return true
     }
-
+    
     override class func canonicalRequest(for request: URLRequest) -> URLRequest {
         return request
     }
-
+    
     override func startLoading() {
         guard let handler = MockURLProtocol.requestHandler else {
             fatalError("MockURLProtocol.requestHandler is not set.")
         }
-
+        
         do {
             let (response, data) = try handler(request)
-
-            if let data = data {
-                client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
+            
+            if let data {
                 client?.urlProtocol(self, didLoad: data)
             }
-
+            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
             client?.urlProtocolDidFinishLoading(self)
         } catch {
             client?.urlProtocol(self, didFailWithError: error)
         }
     }
-
+    
     override func stopLoading() {}
 }
